Regress to running one pipeline step, substep
substep may not apply for all steps, so whereever applicable

this was made necessary by the need to interlink analyses within
the pipeline so that data computed for one analysis can be passed
to the computation of another downstream analysis.

Not that we cannot think of a way to run multi step launch.
We sure can but have not done so for two reasons:
1. We have never run a multi step pipeline launch --- never even tested
2. Parallelization is a pain already for single step launches --- needing calibration
3. The multi-step launch design got us to make implementations that used the config
   to decide what analyses to run --- the simplest implementation would be to run
   all of them in the order established in the pipeline definition. However this led
   to several problems
   3a. What to do with a failed step --- we steps 0 through X - 1 have run, but X fails?
       On a relaunch the pipeline should start with X. With everything in the config approach,
       we will need to change the config.
   3b. While we noticed this problem, nevertheless we decided to move on to explore further
       the topological analysis pipeline space. We had a hypbrid process, keeping our options open,
       and implementing an interface that led to a mismatch between the configured parameters and
       the argued ones when we are running the pipeline for a single step provided at CLI,
       while configuring all the steps that have been pre-computed before the launch of
       the argued step.

The solution has been to require that the step to run be read from the CLI,
along with a substep if the step has substeps (each individual randomization or analysis
is a substep). We can write master script to run each step individually --- in a chain
like circuit building, running steps in parallel whereever possible (independent analyses
can be run) --- serially when upstream computed data is required --- like analyses can
be done only after adjacencies have been computed.

What is the role of the config in all this?
The config is like a manifesto of what the pipeline should do.
TODO: consider renaming the config to a manifesto.
How it is done should be outside the scope the manifesto, and has been done so.
And, when the pipeline run has fully carried out the manifesto,
the config can be used to load the data-store for further analyses.