call an already computed analyses implicitly from another
this will allow the pre computed results to be loaded from the
HDF store and savem recomputation.
The user need not do anything more than naming function arguments properly:

To linke analysis A to analysis B, then the function that computes
B must take an argument with the value of analysis A which may default to None.
If None, then the function implementation will need to worry about how to
compute the value of analysis A. Within the pipeline the value is retrieved
from the the HDF store (if available) and passed to the computation of analysis B.
So not much to worry about -- no extra configuration, just cleaner coding,
and a run of the pipeline in the correct order : A should be run before B.