implement application of connectivity controls to analyses
Consider the example of distance dependent order-2 model for circuit connectivity.
Parameters for the order-2 dd-model are computed as analysis `model_params_dd2`,
which are then passed to the algorithm that implements the control's randomization.
A first run of the pipeline sets up the computation across a configured number of
compute-nodes. Launch of individual computations must be done manually, as is the
later collection of results into the output HDF-store.